Pruebas - SaveFile UseCase Custom values

+++ src/domain/use-cases/save-file.use-case.test.ts
@@ -3,8 +3,21 @@ import {SaveFile} from './save-file.use-case'
 
 describe('SaveFileUseCase', () => {
 
+    const customOptions = {
+        fileContent: 'custom content',
+        fileDestination: 'custom-outputs/file-destination',
+        fileName: 'custom table name',
+    }  
+
+    const customFilePath = `${customOptions.fileDestination}/${customOptions.fileName}.txt`;
+
     afterEach(() => {
-        fs.rmSync('outputs', { recursive: true});
+        const outputFolderExists = fs.existsSync('outputs');
+        if(outputFolderExists) fs.rmSync('outputs', { recursive: true});
+
+        const customOutputFolderExists = fs.existsSync(customOptions.fileDestination);
+        if(customOutputFolderExists) fs.rmSync(customOptions.fileDestination, { recursive: true});
+
     });
     
     test('should save file with default values', () => {
@@ -28,6 +41,21 @@ describe('SaveFileUseCase', () => {
 
 
     });
+
+    test('should save file with custom values', () => {
+        
+        const saveFile = new SaveFile();
+
+        const result = saveFile.execute(customOptions);
+        const fileExists = fs.existsSync(customFilePath); 
+        const fileContent = fs.readFileSync(customFilePath, { encoding: 'utf8' }); // if exists check
+
+        expect(result).toBe(true);
+        expect(fileExists).toBe(true);
+        expect(fileContent).toBe(customOptions.fileContent); 
+
+    });
+    
         
 
 });
